+++ src/main/java/com/filmdoms/community/account/data/constant/AccountStatus.java
@@ -1,4 +1,4 @@
-package com.filmdoms.community.account.data.constants;
+package com.filmdoms.community.account.data.constant;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
