+++ src/main/java/com/filmdoms/community/account/data/constant/OAuthType.java
@@ -1,0 +1,6 @@
+package com.filmdoms.community.account.data.constant;
+
+public enum OAuthType {
+    LOGIN,
+    SIGNUP
+}
