--- src/main/java/com/filmdoms/community/account/data/constants/OAuthType.java
@@ -1,6 +1,0 @@
-package com.filmdoms.community.account.data.constants;
-
-public enum OAuthType {
-    LOGIN,
-    SIGNUP
-}
